Phase 1: extract platform/desktop.js (file-watch + watch toggle + IPC + session)
The Electron integration moves out of main.js into platform/desktop.js:
file-watch reference counting, the watch-toggle UI, session persistence, and
the main-process IPC wiring (file-open / file-changed / close-tab, native
menus, custom CSS). It imports tabs.js (createTab/closeTab/renderTabBar) — a
one-way dependency, no cycle — and receives renderMarkdown via configureDesktop
(named reloadDoc to avoid the harness global collision with tabs.js's render
callback). main.js wires configureTabs with the desktop watch/session funcs.

main.js: 992 -> 588 lines. Pure refactor: build + lint + 297 tests green.
